text_analyzer/requirements.txt
astor==0.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
conductor-python==1.1.10
dacite==1.8.1
idna==3.10
prometheus_client==0.21.1
shortuuid==1.0.13
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
requests
PyPDF2
openai
